\name{ggkaboom-package}
\alias{ggkaboom-package}
\alias{ggkaboom}
\docType{package}
\title{
\packageTitle{ggkaboom}
}
\description{
\packageDescription{ggkaboom}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ggkaboom}
\packageIndices{ggkaboom}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ggkaboom}

Maintainer: \packageMaintainer{ggkaboom}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
